refactor(react-doctor): decompõe SchemaEditor em header/banners/dialogs (#354)
SchemaEditor sai do no-giant-component (441 → ~260 linhas), extrações
locais sem mudança de comportamento:

- SchemaEditorHeader: toggle Visual/Código + versão + contadores + ações
- SchemaEditorBanners: ajuda de versionamento + recuperação de legado
- SchemaEditorDialogs: AlertDialogs de backfill e Publicar MAJOR
- useSchemaEditorDialogs movido para arquivo próprio co-localizado
  (leva junto VERSIONING_HELP_KEY)

Estado (mode/fields/code/guiErrors/isPending), handlers e o miolo
GUI-vs-Monaco permanecem no pai — cruzam mode/isPending.
